modified:   MetaTradeNode/BlockchainService.h 	modified:   MetaTradeNode/LevelDBLocalImpl.cpp 	modified:   MetaTradeNode/LevelDBLocalImpl.h 	modified:   MetaTradeNode/LocalService.h 	modified:   MetaTradeNode/MetaTradeBlockchainImpl.cpp 	modified:   MetaTradeNode/MetaTradeBlockchainImpl.h 	modified:   MetaTradeNode/MetaTradeBlockchainModels.cpp 	modified:   MetaTradeNode/MetaTradeBlockchainModels.h 	modified:   MetaTradeNode/MetaTradeNode.cpp 	modified:   MetaTradeNode/MetaTradeNode.h 	modified:   MetaTradeNode/example.cpp 	modified:   MetaTradeNode/include/webstomp++/WebStompClient.h 	modified:   MetaTradeNode/include/webstomp++/WebStompType.h 	modified:   MetaTradeNode/lib/libWebStomp++.lib 	modified:   dlls/libWebStomp++.dll 	new file:   shared/Debug/libMetaTradeNode_d.dll 	new file:   shared/Debug/libMetaTradeNode_d.lib 	new file:   shared/Release/libMetaTradeNode.dll 	new file:   shared/Release/libMetaTradeNode.lib

+++ MetaTradeNode/MetaTradeBlockchainModels.cpp
@@ -40,8 +40,8 @@ void metatradenode::Block::calMerkleHash() {
     this->merkle_hash = hashTree.front();
 }
 
-long metatradenode::Block::getBlockCommision(){
-    long sum = 0;
+long long metatradenode::Block::getBlockCommision(){
+    long long sum = 0;
     for (auto& trade : block_body) {
         sum += trade.commission;
     }
