+++ MetaTradeNode/MetaTradeNode.h
@@ -32,8 +32,8 @@ namespace metatradenode {
 		void init();
 		void reload();
 		void run(bool sync = true);
-		long queryAmount(const char* address, const char* item_id);
-		void submitTrade(const char* receiver, const char* item_id, long amount);
+		long long queryAmount(const char* address, const char* item_id);
+		void submitTrade(const char* receiver, const char* item_id, long long amount);
 	};
 }
